Tracker interface, refactors
