+++ README.md
@@ -1,1 +1,6 @@
-"# SocialDistanceEstimation" 
+# SocialDistanceEstimation
+
+_TODO: description_
+
+## Installation
+
